feat: ロジックをカスタムフックに移す
